[#10] Move email handler to separate file
